fix mvc structure
