+++ public/profile.html
@@ -197,7 +197,7 @@
                                 </div>
                             </div>
                             <div class="col-md-6">
-                                <form method="post" action="/api/1.0/uploadShares" enctype="multipart/form-data" class="m-top-20" id="comment-form" style="display:none">
+                                <form method="post" action="/api/1.0/sharing" enctype="multipart/form-data" class="m-top-20" id="comment-form" style="display:none">
                                     <input type="file" name="main_image" placeholder="約會主題照" class="form-control" required="">
                                     <textarea class="form-control m-top-20" name="story" rows="6" placeholder="向大家分享你的約會心得吧" required=""></textarea>
                                     <input class="btn btn-primary m-top-20" type="submit" value="分享行程" />
