+++ test/user_test.js
@@ -1,4 +1,4 @@
-const { dbStatus } = require('../routes/db');
+const { dbStatus } = require('../server/db');
 const { app } = require('../app');
 const chai = require('chai');
 const chaiHttp = require('chai-http');
